// apps/api/src/modules/member/dto/update-member.dto.ts
export class UpdateMemberDto {
  name?: string;
  email?: string;
  siteId?: string;
  employeeNumber?: string;
  gender?: string;
  dateOfBirth?: string;
  disabilityType?: string;
  disabilityGrade?: string;
  handbookType?: string;
  handbookIssuedAt?: string;
  handbookExpiresAt?: string;
  employmentType?: string;
  enrolledAt?: string;
  status?: string;
  workExperience?: string;
  preferredWorkAreas?: string;
  avatarId?: string;
}